+++ lib/QECCore/src/QECCore.jl
@@ -3,35 +3,48 @@ module QECCore
 using SparseArrays
 using LinearAlgebra
 using Combinatorics
+using Graphs
+
+using DocStringExtensions
 
 # interfaces
-export distance, parity_matrix, code_n, code_s, code_k, parity_matrix_x, parity_matrix_z, rate, metacheck_matrix_x, metacheck_matrix_z, metacheck_matrix
+export distance, parity_matrix, code_n, code_s, code_k, parity_matrix_x, parity_matrix_z, parity_matrix_xz, rate, metacheck_matrix_x, metacheck_matrix_z, metacheck_matrix
 export AbstractECC, AbstractQECC, AbstractCECC, AbstractCSSCode, AbstractDistanceAlg
 
 # QEC Codes
-export Perfect5, Cleve8
+export Perfect5, Cleve8, Gottesman
 
 # CSS Codes
-export Toric, Bitflip3, Phaseflip3, Shor9, Steane7, Surface, CSS, QuantumReedMuller
+export Toric, Bitflip3, Phaseflip3, Shor9, Steane7, Surface, CSS, QuantumReedMuller, Triangular488, Triangular666, QuantumTannerGraphProduct, CyclicQuantumTannerGraphProduct
 
 # Classical Codes
-export RepCode, ReedMuller, RecursiveReedMuller
+export RepCode, ReedMuller, RecursiveReedMuller, Golay, Hamming
 
 include("interface.jl")
 include("codes/util.jl")
-include("codes/css.jl")
-include("codes/fivequbit.jl")
-include("codes/reptetion.jl")
-include("codes/toric.jl")
-include("codes/clevecode.jl")
-include("codes/shorcode.jl")
-include("codes/steanecode.jl")
-include("codes/surface.jl")
-include("codes/bitflipcode.jl")
 
-include("codes/reedmuller.jl")
-include("codes/recursivereedmuller.jl")
-include("codes/quantumreedmuller.jl")
+# Classical Codes
+include("codes/classical/hamming.jl")
+include("codes/classical/reptetion.jl")
+include("codes/classical/golay.jl")
+
+# Quantum Codes
+include("codes/quantum/css.jl")
+include("codes/quantum/fivequbit.jl")
+include("codes/quantum/toric.jl")
+include("codes/quantum/clevecode.jl")
+include("codes/quantum/shorcode.jl")
+include("codes/quantum/steanecode.jl")
+include("codes/quantum/surface.jl")
+include("codes/quantum/bitflipcode.jl")
+include("codes/quantum/gottesman.jl")
+include("codes/quantum/color_codes.jl")
+include("codes/quantum/quantumtannergraphproduct.jl")
+
+# Reed-Muller Codes
+include("codes/classical/reedmuller.jl")
+include("codes/classical/recursivereedmuller.jl")
+include("codes/quantum/quantumreedmuller.jl")
 
 function __init__()
     if isdefined(Base.Experimental, :register_error_hint)