+++ lib/QECCore/test/codes/quantumtannergraphproduct.jl
@@ -1,17 +1,17 @@
-@testitem "ECC Quantum Expander codes" begin
+@testitem "Quantum Tanner Graph Codes" begin
 
-    using SparseArrays
     using Random
-    using LinearAlgebra
     using Graphs
     using Graphs: is_bipartite
+    using SparseArrays
     using QuantumClifford
     using QuantumClifford: stab_looks_good
     using QuantumClifford.ECC
-    using QuantumClifford.ECC: parity_checks_xz, tanner_graph_from_parity_matrix, parity_matrix_from_tanner_graph
-    using QuantumClifford.ECC: Golay, RepCode, ReedMuller
+    using QuantumClifford.ECC: parity_matrix_x, parity_matrix_z
+    using QECCore
+    using QECCore: tanner_graph_from_parity_matrix, parity_matrix_from_tanner_graph
 
-    function verify_orthogonality(HX::SparseMatrixCSC{Bool,Int}, HZ::SparseMatrixCSC{Bool,Int})
+    function verify_orthogonality(HX::AbstractMatrix, HZ::AbstractMatrix)
         product_mod2 = mod.(Int.(HX) * transpose(Int.(HZ)), 2)
         return all(product_mod2 .== 0)
     end
@@ -29,16 +29,16 @@
     function generate_parity_checks(code_type::Symbol, args...)
         if code_type == :RepCode
             n = args[1]
-            return parity_checks(RepCode(n))
+            return parity_matrix(RepCode(n))
         elseif code_type == :ReedMuller
             r, m = args
-            return sparse(parity_checks(ReedMuller(r, m)))
+            return ReedMuller(r, m)
         elseif code_type == :Golay
             n = args[1]
             if n == 23
-                return sparse(Matrix{Bool}(parity_checks(Golay(n))))
+                return parity_matrix(Golay(n))
             elseif n == 24
-                return sparse(Matrix{Bool}(parity_checks(Golay(n))))
+                return parity_matrix(Golay(n))
             else
                 error("Golay code only supports n = 23 or 24")
             end
@@ -50,11 +50,11 @@
     @testset "4.1: Validity of the Construction of Q(G1 × G2)" begin
         @testset "Repetition codes" begin
             for n in 3:20
-                H = sparse(parity_checks(RepCode(n)))
+                H = parity_matrix(RepCode(n))
                 c = QuantumTannerGraphProduct(H, H)
                 @test stab_looks_good(parity_checks(c); remove_redundant_rows=true)
-                hx, hz = parity_checks_xz(c)
-                @test verify_orthogonality(sparse(hx), sparse(hz))
+                hx, hz = parity_matrix_x(c), parity_matrix_z(c)
+                @test verify_orthogonality(hx, hz)
             end
         end
 
@@ -62,28 +62,28 @@
             for n in 3:20
                 c = CyclicQuantumTannerGraphProduct(n)
                 @test stab_looks_good(parity_checks(c); remove_redundant_rows=true)
-                hx, hz = parity_checks_xz(c)
-                @test verify_orthogonality(sparse(hx), sparse(hz))
+                hx, hz = parity_matrix_x(c), parity_matrix_z(c)
+                @test verify_orthogonality(hx, hz)
             end
         end
 
         @testset "Golay codes" begin
             for n in [23, 24]
-                H = sparse(Matrix{Bool}(parity_checks(Golay(n))))
+                H = parity_matrix(Golay(n))
                 c = QuantumTannerGraphProduct(H, H)
                 @test stab_looks_good(parity_checks(c); remove_redundant_rows=true)
-                hx, hz = parity_checks_xz(c)
-                @test verify_orthogonality(sparse(hx), sparse(hz))
+                hx, hz = parity_matrix_x(c), parity_matrix_z(c)
+                @test verify_orthogonality(hx, hz)
             end
         end
 
         @testset "Reed-Muller codes" begin
             for m in 3:5, r in 1:m-1
-                H = sparse(parity_checks(ReedMuller(r, m)))
+                H = parity_matrix(ReedMuller(r, m))
                 c = QuantumTannerGraphProduct(H, H)
                 @test stab_looks_good(parity_checks(c); remove_redundant_rows=true)
-                hx, hz = parity_checks_xz(c)
-                @test verify_orthogonality(sparse(hx), sparse(hz))
+                hx, hz = parity_matrix_x(c), parity_matrix_z(c)
+                @test verify_orthogonality(hx, hz)
             end
         end
     end